# src/main/resources/swagger_config/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/v1/reservation:
    get:
      tags:
      - Reservation
      operationId: getAllReservations
      parameters:
      - name: status
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - PENDING
          - APPROVED
          - DECLINED
          - CANCELLED
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseListReservationResponse'
    put:
      tags:
      - Reservation
      operationId: updateReservationByUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationByAdmin'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseReservationResponse'
    post:
      tags:
      - Reservation
      operationId: createReservation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseReservationResponse'
...
  /api/v1/employee:
    get:
      tags:
      - Employee
      operationId: getAllEmployee
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseListEmployeeResponse'
    put:
      tags:
      - Employee
      operationId: updateEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseEmployeeResponse'
...
  /api/v1/reservation/{id}:
    get:
      tags:
      - Reservation
      operationId: getReservationById
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseReservationResponse'
  /api/v1/reservation/success:
    get:
      tags:
      - Reservation
      operationId: success
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/reservation/status/{id}:
    get:
      tags:
      - Reservation
      operationId: updateStatusBySendMail
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: action
        in: query
        required: true
        style: form
        explode: true
        schema:
          type: string
          enum:
          - PENDING
          - APPROVED
          - DECLINED
          - CANCELLED
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "200":
          description: OK
  /api/v1/reservation/employee:
    get:
      tags:
      - Reservation
      operationId: getAllByEmployee
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseListReservationResponse'
  /api/v1/reservation/download:
    get:
      tags:
      - Reservation
      operationId: downloadReservation
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "200":
          description: OK
  /api/v1/reservation/available-room:
    get:
      tags:
      - Reservation
      operationId: getReservationStatus
      parameters:
      - name: request
        in: query
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/InsertDateRequest'
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseListGetReservationStatusResponse'
...
components:
  schemas:
    CommonResponseString:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: string
    CommonResponseErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: integer
          format: int64
...
    RoomFacilityResponse:
      type: object
      properties:
        roomId:
          type: string
        roomFacilityName:
          type: string
...
    RoomResponse:
      type: object
      properties:
        roomId:
          type: string
        roomType:
          type: string
        roomCapacity:
          type: integer
          format: int32
        roomFacilities:
          type: array
          items:
            $ref: '#/components/schemas/RoomFacilityResponse'
    UpdateReservationByAdmin:
      required:
      - endTime
      - idReservation
      - roomId
      - startTime
      type: object
      properties:
        startTime:
          pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
          type: string
        endTime:
          pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
          type: string
        roomId:
          type: string
        idReservation:
          type: string
    CommonResponseReservationResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/ReservationResponse'
    EmployeeResponse:
      type: object
      properties:
        employeeId:
          type: string
        employeeName:
          type: string
        department:
          type: string
    EquipmentResponse:
      type: object
      properties:
        equipmentId:
          type: string
        equipmentName:
          type: string
    ReservationResponse:
      type: object
      properties:
        reservationId:
          type: string
        employee:
          $ref: '#/components/schemas/EmployeeResponse'
        room:
          $ref: '#/components/schemas/RoomResponse'
        reserveDate:
          type: string
          format: date
        startTime:
          type: string
          format: date
        endTime:
          type: string
          format: date
        reservationStatus:
          type: string
          enum:
          - PENDING
          - APPROVED
          - DECLINED
          - CANCELLED
        reservationDescription:
          type: string
        equipments:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentResponse'
    EquipmentRequest:
      type: object
      properties:
        idEquipment:
          type: string
        equipmentName:
          type: string
...
    UpdateEmployeeRequest:
      type: object
      properties:
        employeeId:
          type: string
        employeeName:
          type: string
        department:
          type: string
        phoneNumber:
          type: string
        corporateEmail:
          type: string
    CommonResponseEmployeeResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/EmployeeResponse'
    ReservationRequest:
      required:
      - employeeId
      - endTime
      - reservationDescription
      - roomId
      - startTime
      type: object
      properties:
        startTime:
          pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
          type: string
        endTime:
          pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
          type: string
        reservationDescription:
          type: string
        roomId:
          type: string
        employeeId:
          type: string
        equipmentRequests:
          type: array
          items:
            $ref: '#/components/schemas/EquipmentRequest'
...
    CommonResponseListReservationResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReservationResponse'
    InsertDateRequest:
      required:
      - date
      type: object
      properties:
        date:
          pattern: "^\\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$"
          type: string
    CommonResponseListGetReservationStatusResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/GetReservationStatusResponse'
    GetReservationStatusResponse:
      type: object
      properties:
        reservationId:
          type: string
        employee:
          type: string
        room:
          type: string
        equipments:
          type: array
          items:
            type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    CommonResponseListEmployeeResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeResponse'
    inline_response_400:
      oneOf:
      - $ref: '#/components/schemas/CommonResponseErrorResponse'
      - $ref: '#/components/schemas/CommonResponseString'